+++ src/pieces/piece.rs
@@ -1,7 +1,7 @@
 pub trait Piece: Send + Sync {
-    fn get_pos(&self) -> (usize, usize);
-    fn move_piece(&mut self, new_row: usize, new_col: usize, board: &mut Vec<Vec<&'static str>>);
-    fn get_valid_moves(&self, board: &Vec<Vec<&'static str>>) -> Vec<(usize, usize)>;
+    fn get_pos(&self) -> (i8, i8);
+    fn move_piece(&mut self, new_row: i8, new_col: i8, board: &mut Vec<Vec<&'static str>>);
+    fn get_valid_moves(&self, board: &Vec<Vec<&'static str>>) -> Vec<(i8, i8)>;
     fn get_color(&self) -> char;
     fn get_code(&self) -> &'static str;
 }
